Add files via upload
This notebook examines the mechanism of gradient descent and its dependence on key parameter like the number of iterations and the learning rate. The goal is to understand how gradient descent works and to gain hands-on experience with TensorFlow.

I begin with a simple case: a one-dimensional function. An interactive widget lets users adjust parameters and visually observe the gradient descent path. This provides an intuitive understanding of how the algorithm converges.

Next, I extend the exploration to two dimensions using TensorFlow. The gradient descent path is plotted on a three-dimensional surface representing the function. The widget shows how the optimization process works on more than one dimension.

Finally, I present another interactive widget to visualize the gradient descent path on the two-dimensional level curves of the function, offering a complementary perspective on the optimization process.

Through these exercises, the notebook aims to strengthen understanding of both the gradient descent algorithm and the fundamentals of TensorFlow.

Part of the code in section 2 and 3 is an adaptation of the code presented in file Utah_Gradient (see references).